clocks-8960: gpu: allow oc to 487.5 & 533.333 Mhz
In the new version of this file QCOM tried to merge all apq8064 freq tables.
We can't do that.
APQ8064aa & APQ8064 share a pll freq of 975Mhz, but APQ8064ab has only 900Mhz
on PLL15! So if we were to share one freq table, then the 450Mhz entry would in
fact be 487.5Mhz.
So we just create a dedicated table for apq8064&aa and use that.

PS: It is probable that QCOM did this because apq8064&aa are limited to 400Mhz anyway
and don't use the pll15 anywhere within their limits. However, this sucks for us now.

Signed-off-by: Dennis Rassmann <showp1984@gmail.com>